% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expit}
\alias{expit}
\alias{nexpit}
\title{Calculate expit (inverse logit) of \code{phi}}
\usage{
expit(phi)

nexpit(phi)
}
\arguments{
\item{phi}{}
}
\description{
Calculate expit (inverse logit) of \code{phi}
}
\section{Functions}{
\itemize{
\item \code{nexpit}: Nexpit of \code{phi} for inverse Hessian
}}

